requirements.txt
PyYAML==6.0
connect-markdown-renderer==3.0.0
PyMuPDF==1.23.8
